refactor: RestaurantBatchService 생성 및 CategoryDictionary -> CategoryType 이름 변경
