+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/restaurant/service/RestaurantService.java
@@ -2,27 +2,22 @@ package com.jjangiji.hankkimoa.restaurant.service;
 
 import com.jjangiji.hankkimoa.common.exception.ExceptionCode;
 import com.jjangiji.hankkimoa.common.exception.HankkiMoaException;
-import com.jjangiji.hankkimoa.restaurant.domain.Address;
-import com.jjangiji.hankkimoa.restaurant.domain.Category;
 import com.jjangiji.hankkimoa.restaurant.domain.Menu;
 import com.jjangiji.hankkimoa.restaurant.domain.OpeningHour;
 import com.jjangiji.hankkimoa.restaurant.domain.RecommendationFeedback;
 import com.jjangiji.hankkimoa.restaurant.domain.Restaurant;
 import com.jjangiji.hankkimoa.restaurant.domain.RestaurantImage;
-import com.jjangiji.hankkimoa.restaurant.repository.CategoryRepository;
 import com.jjangiji.hankkimoa.restaurant.repository.MenuRepository;
 import com.jjangiji.hankkimoa.restaurant.repository.OpeningHoursRepository;
 import com.jjangiji.hankkimoa.restaurant.repository.RecommendationFeedbackRepository;
 import com.jjangiji.hankkimoa.restaurant.repository.RestaurantImageRepository;
 import com.jjangiji.hankkimoa.restaurant.repository.RestaurantRepository;
 import com.jjangiji.hankkimoa.restaurant.service.dto.reqeust.RecommendationFeedbackRequest;
-import com.jjangiji.hankkimoa.restaurant.service.dto.reqeust.RestaurantCreateRequest;
 import com.jjangiji.hankkimoa.restaurant.service.dto.response.MenuResponse;
 import com.jjangiji.hankkimoa.restaurant.service.dto.response.RecommendServerRestaurantsResponse;
 import com.jjangiji.hankkimoa.restaurant.service.dto.response.RestaurantResponse;
 import com.jjangiji.hankkimoa.restaurant.service.dto.response.RestaurantSearchResponse;
 import com.jjangiji.hankkimoa.restaurant.service.dto.response.RestaurantSimpleResponse;
-import com.jjangiji.hankkimoa.restaurant.util.CategoryMapper;
 import com.jjangiji.hankkimoa.restaurant.util.DayOfWeekMapper;
 import com.jjangiji.hankkimoa.user.domain.User;
 import com.jjangiji.hankkimoa.user.repository.BookmarkRepository;
@@ -34,7 +29,6 @@ import java.time.LocalDate;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 @RequiredArgsConstructor
 @Service
@@ -46,68 +40,7 @@ public class RestaurantService {
     private final RestaurantImageRepository restaurantImageRepository;
     private final OpeningHoursRepository openingHoursRepository;
     private final MenuRepository menuRepository;
-    private final CategoryRepository categoryRepository;
     private final BookmarkRepository bookmarkRepository;
-
-    @Transactional
-    public void createRestaurants(List<RestaurantCreateRequest> requests) {
-        Set<String> uniqueIds = restaurantRepository.findAllUniqueId();
-        CategoryMapper categoryMapper = new CategoryMapper(categoryRepository.findAll());
-
-        for (RestaurantCreateRequest request : requests) {
-            if (request.menus() == null) continue;
-            if (request.id() == null || uniqueIds.contains(request.id())) continue;
-
-            Restaurant restaurant = saveRestaurant(request, categoryMapper);
-            uniqueIds.add(restaurant.getUniqueId());
-            saveRestaurantImages(restaurant, request);
-            saveOpeningHours(restaurant, request);
-            saveMenus(restaurant, request);
-        }
-    }
-
-    private Restaurant saveRestaurant(RestaurantCreateRequest request, CategoryMapper categoryMapper) {
-        Category category = null;
-        Address address = null;
-        if (request.category() != null) category = categoryMapper.mapByKeyword(request.category());
-        if (request.address() != null) address = new Address(0, 0, request.address());
-
-        Restaurant restaurant = new Restaurant(category, request.name(), request.id(), request.menu_average(), address);
-        return restaurantRepository.save(restaurant);
-    }
-
-    private void saveRestaurantImages(Restaurant restaurant, RestaurantCreateRequest request) {
-        if (request.images() == null) return;
-
-        List<RestaurantImage> restaurantImages = request.images()
-                .stream()
-                .map(image -> new RestaurantImage(restaurant, image))
-                .toList();
-        restaurantImageRepository.saveAll(restaurantImages);
-    }
-
-    private void saveOpeningHours(Restaurant restaurant, RestaurantCreateRequest request) {
-        if (request.openingHours() == null) return;
-
-        List<OpeningHour> openingHours = request.openingHours().stream()
-                .map(openingHour -> new OpeningHour(restaurant, openingHour.dayOfWeek(),
-                        openingHour.hours().startTime(), openingHour.hours().endTime(),
-                        openingHour.hours().breakStartTime(), openingHour.hours().breakEndTime(),
-                        openingHour.hours().lastOrderTime()))
-                .toList();
-        openingHoursRepository.saveAll(openingHours);
-    }
-
-    private void saveMenus(Restaurant restaurant, RestaurantCreateRequest request) {
-        if (request.menus() == null) return;
-
-        List<Menu> menus = request.menus().stream()
-                .map(menu -> new Menu(restaurant, menu.name(),
-                        menu.price(), menu.imgUrl(),
-                        menu.isMain(), menu.introduce()))
-                .toList();
-        menuRepository.saveAll(menus);
-    }
 
     @Transactional
     public Long createRecommendationFeedback(User user, RecommendationFeedbackRequest request) {
@@ -169,7 +102,7 @@ public class RestaurantService {
 
     private String formatOpeningHour(Optional<OpeningHour> openingHour) {
         return openingHour
-                .map(oh -> String.format("%s %s-%s", oh.getDayOfWeek(), oh.getOpenTime(), oh.getCloseTime()))
+                .map(oh -> String.format("%s %s - %s", oh.getDayOfWeek(), oh.getOpenTime(), oh.getCloseTime()))
                 .orElse(null);
     }
