+++ backend/hankkimoa/src/test/java/com/jjangiji/hankkimoa/expense/service/ExpenseServiceTest.java
@@ -12,7 +12,7 @@ import com.jjangiji.hankkimoa.expense.service.dto.response.CommunityExpenseRespo
 import com.jjangiji.hankkimoa.expense.service.dto.response.TodayExpenses;
 import com.jjangiji.hankkimoa.expense.service.dto.response.MonthlyExpenseResponse;
 import com.jjangiji.hankkimoa.restaurant.domain.Address;
-import com.jjangiji.hankkimoa.restaurant.domain.CategoryDictionary;
+import com.jjangiji.hankkimoa.restaurant.domain.CategoryType;
 import com.jjangiji.hankkimoa.restaurant.domain.Category;
 import com.jjangiji.hankkimoa.restaurant.domain.Restaurant;
 import com.jjangiji.hankkimoa.restaurant.repository.CategoryRepository;
@@ -55,7 +55,7 @@ class ExpenseServiceTest extends IntegrationTest {
     @BeforeEach
     void setUp() {
         user = userRepository.save(new User("hankkimoa@gmail.com", "한끼", "hankkiImage", LoginType.KAKAO, Role.USER));
-        Category category = categoryRepository.save(new Category(CategoryDictionary.한식));
+        Category category = categoryRepository.save(new Category(CategoryType.한식));
         restaurant  = restaurantRepository.save(new Restaurant(category, "한끼식당", "12345", 10000, address));
         expenseSavingGoal = expenseSavingGoalRepository.save(new ExpenseSavingGoal(user, 70_000, now, sevenDayAfter));
     }
