+++ backend/hankkimoa/src/test/java/com/jjangiji/hankkimoa/restaurant/repository/RestaurantRepositoryTest.java
@@ -6,7 +6,7 @@ import com.jjangiji.hankkimoa.expense.domain.ExpenseSavingGoal;
 import com.jjangiji.hankkimoa.expense.repository.ExpenseRepository;
 import com.jjangiji.hankkimoa.expense.repository.ExpenseSavingGoalRepository;
 import com.jjangiji.hankkimoa.restaurant.domain.Category;
-import com.jjangiji.hankkimoa.restaurant.domain.CategoryDictionary;
+import com.jjangiji.hankkimoa.restaurant.domain.CategoryType;
 import com.jjangiji.hankkimoa.restaurant.domain.Restaurant;
 import com.jjangiji.hankkimoa.restaurant.service.dto.response.RestaurantAverageRatingResponse;
 import com.jjangiji.hankkimoa.user.domain.LoginType;
@@ -41,7 +41,7 @@ class RestaurantRepositoryTest extends RepositoryTest {
         User user = userRepository.save(new User("hankkimoa@gmail.com", "한끼", "hankkiImage", LoginType.KAKAO, Role.USER));
         ExpenseSavingGoal expenseSavingGoal = expenseSavingGoalRepository.save(expenseSavingGoalRepository.save(
                 new ExpenseSavingGoal(user, 70_000, LocalDate.now(), LocalDate.now().plusDays(6))));
-        Category category = categoryRepository.save(new Category(CategoryDictionary.한식));
+        Category category = categoryRepository.save(new Category(CategoryType.한식));
         Restaurant restaurant  = restaurantRepository.save(new Restaurant(category, "한끼식당", expectedUniqueId, 10000, null));
         Expense expense = expenseRepository.save(new Expense(expenseSavingGoal, restaurant,
                 "한끼식당", "순두부", 8_000,
