+++ backend/hankkimoa/src/test/java/com/jjangiji/hankkimoa/user/service/UserServiceTest.java
@@ -4,7 +4,7 @@ import com.jjangiji.hankkimoa.common.exception.ExceptionCode;
 import com.jjangiji.hankkimoa.common.exception.HankkiMoaException;
 import com.jjangiji.hankkimoa.config.IntegrationTest;
 import com.jjangiji.hankkimoa.restaurant.domain.Category;
-import com.jjangiji.hankkimoa.restaurant.domain.CategoryDictionary;
+import com.jjangiji.hankkimoa.restaurant.domain.CategoryType;
 import com.jjangiji.hankkimoa.restaurant.repository.CategoryRepository;
 import com.jjangiji.hankkimoa.user.domain.LoginType;
 import com.jjangiji.hankkimoa.user.domain.Role;
@@ -44,8 +44,8 @@ class UserServiceTest extends IntegrationTest {
     @Test
     void updateCategories() {
         // given
-        Category category1 = categoryRepository.save(new Category(CategoryDictionary.한식));
-        Category category2 = categoryRepository.save(new Category(CategoryDictionary.양식));
+        Category category1 = categoryRepository.save(new Category(CategoryType.한식));
+        Category category2 = categoryRepository.save(new Category(CategoryType.양식));
         userCategoryRepository.save(new UserCategory(user, category1));
 
         // when
